requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0.0

# LangChain ecosystem (1.0+ versions)
langgraph>=1.0.0
langchain-core>=1.0.0
langchain-openai>=1.0.0

# Research APIs
tavily-python>=0.5.0
requests>=2.31.0
duckduckgo-search>=6.0.0

# Web Demo
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
streamlit>=1.39.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
